Move util.py to its own package, and begin refactoring.
RefKeyDictionary moved to its own file.
This will be done for other classes, functions, etc., and
unit tests will be done for each.
